IGNITE-28733 Move QueryEntityPatch to internal
